x86: Name parameters in function declarations
With the wider testing, some more violations have been spotted.  This
addresses violations of Rule 8.2 (parameters must be named).

livepatch_op() has a further problem.  The other declaration (the stub for
!LIVEPATCH builds) has a parameter named op, but the definition has the
parameter named livepatch.  Rename it to op which is more concise and better
in line with it's type name.

No functional change.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Nicola Vetrini <nicola.vetrini@bugseng.com>