xen/gen_hypercall: Fix if indent
On x86, the opening if of call_handlers_*() is indented 4 spaces which
breaks alignment.  The alignment is broken by printf("    "), but
that actually aligns ARM's opening "if ( likely() )".  Move the indent
to the if that needs it.

This results in x86 hypercall-defs.h changes like:

 #define call_handlers_hvm64(num, ret, a1, a2, a3, a4, a5) \
 ({ \
     uint64_t mask = (num) > 63 ? 0 : 1ULL << (num); \
-        if ( likely(mask & ((1ULL << __HYPERVISOR_vcpu_op) | (1ULL << __HYPERVISOR_event_channel_op))) ) \
+    if ( likely(mask & ((1ULL << __HYPERVISOR_vcpu_op) | (1ULL << __HYPERVISOR_event_channel_op))) ) \
     { \
         if ( (num) == __HYPERVISOR_vcpu_op ) \

ARM's hypercall-defs.h does not change.

Signed-off-by: Jason Andryuk <jason.andryuk@amd.com>
Reviewed-by: Jan Beulich <jbeulich@suse.com>